+++ src/main/java/com/realive/dto/logs/PenaltyLogCreateRequest.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class PenaltyLogCreateRequest {
-    private Integer customerId;
+    private Long customerId; // Integer → Long 변경
     private String reason;
     private Integer points;
     private String description;
